add new_slot_execution_outputs_server stream
